Combine multiset files
